% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_station.R
\name{get_best_station}
\alias{get_best_station}
\title{get best station}
\usage{
get_best_station(stations_in_radius)
}
\arguments{
\item{stations_in_radius}{Dataframe with all the gas stations with all the gas stations in distance given by the user}
}
\value{
A one row \code{dataframe} containing the data of the gas station with the cheapest price
}
\description{
picks the cheapest gas station from all the gas stations in the distance given by the user
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
