% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_data.R
\name{get_map_data}
\alias{get_map_data}
\title{get map data}
\usage{
get_map_data(stations_in_radius, lat, long)
}
\arguments{
\item{stations_in_radius}{Data of the gas station in the radius}

\item{lat}{Latitude of the user}

\item{long}{Longitude of the user}
}
\value{
A \code{dataframe} with information about the gas stations and the user
}
\description{
creation of map data: take data from gas stations and adds spatial data of user
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
