% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_in_radius.R
\name{get_stations_in_radius}
\alias{get_stations_in_radius}
\title{get stations in radius}
\usage{
get_stations_in_radius(stations_data, max_dist)
}
\arguments{
\item{stations_data}{A \code{dataframe} with the data of the gas stations}

\item{max_dist}{A numeric value which determines the maximum distance in km}
}
\value{
A \code{dataframe} with the data about the gas stations in the radius
}
\description{
Selects form all the data only those gas stations which are in the radius given by the user
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
