% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions project group c.R
\name{measure_dist}
\alias{measure_dist}
\title{measuring distance function}
\usage{
measure_dist(addresses)
}
\arguments{
\item{addresses}{A \code{vector} with the addresses of the gasstation and the location of the applicant}
}
\value{
A \code{dataframe} with information about:\describe{
   \item{the addresses}
   \item{the latitutes}
   \itemt{the longitute}
   \item{the distances}
}
}
\description{
computes the distances between the current location and the location of the gas stations
}
\examples{
addresses <- c("Unter den Linden 25, 10117 Berlin", "Av. de la Gare 12, 1003 Lausanne", "Kurfürstendamm 100, 10709 Berlin")
measure_dist(addresses = addresses)
}
\author{
Clarence, Franz, Mathieu, Lukas
}
