% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{shiny app}
\usage{
run_app()
}
\value{
A shiny app which can be used to find the best gas station
}
\description{
webpage for finding the best gas station of the user.
The shiny app includes widgets for adjusting the distance in which gas station should be considered
and for the type of fuel the user needs. The app issues the name of the most suitable gas station
and a map with the location of the user and all the gas stations which are in the given distance.
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
