% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{update_data}
\usage{
update_data()
}
\value{
A \code{dataframe} with gas stations data including longitude and latitude
}
\description{
Adds the longitude and latitude data to the web scrapped data
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
