% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrap_data.R
\name{webscrap_data}
\alias{webscrap_data}
\title{webscrap data}
\usage{
webscrap_data(save_csv_folder = here::here("data/gas_station_data.csv"))
}
\value{
A \code{dataframe} with gas stations data including:\describe{
   \item{brand}
   \item{name}
   \itemt{type of fuel}
   \item{price of fuel}
   \item{address}
}
}
\description{
webscrapping of data from fuelo website
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
