% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_dist.R
\name{measure_dist}
\alias{measure_dist}
\title{measuring distance function}
\usage{
measure_dist(data_gasstations, lat, long)
}
\arguments{
\item{data_gasstations}{A \code{dataframe} with all information about the gas stations}

\item{lat}{Latitude of the user}

\item{long}{Longitude of the user}
}
\value{
The gas station \code{dataframe} with an extra distance column
}
\description{
computes the distances between the current location and the location of the gas stations
}
\examples{
"hallo"
}
\author{
Clarence, Franz, Mathieu, Lucas
}
